refactor: move page preview stuff into preview_manager.js
